feat(7160test): 7100 开机 A7 参数读回（降噪/DFBC/WDRC）+ I2C HAL 扩展
- app.h 新增 DSP7100_READBACK_ENABLE（默认开），main() 在 7100 上电 init 后调用
- dsp_7100_cmd.c/h：读回 4 程序 × (0xAE 降噪 / 0x32 DFBC / 0x177 WDRC low+high gain)，
  按参考机全量读抓包帧序列收发（A7 03 37 选中 + A7 01 读块 + 82 收尾），解析存
  g_dsp7100_read[4] 并 UART 打印；发包间隔对齐参考 40-120ms
- i2c_7100_hal read/write 长度扩 uint16（>255B 单事务读，如 0x177=378B）
- i2c_7100_hal 读写超时显式标失败 + Sys_I2C_Reset（原超时误报成功）

联调进行中：能取回数据但块响应头非预期 46，见 docs/7100_参数读回开发记录.md

+++ peripheral_server_sleep7160test/include/i2c_7100_hal.h
@@ -34,10 +34,11 @@ extern "C" {
 bool i2c_7100_hal_init(void);
 
 /* Write len bytes to I2C slave. Returns true on success. */
-bool i2c_7100_write(uint8_t addr, const uint8_t *data, uint8_t len);
+bool i2c_7100_write(uint8_t addr, const uint8_t *data, uint16_t len);
 
-/* Read len bytes from I2C slave. Returns true on success. */
-bool i2c_7100_read(uint8_t addr, uint8_t *data, uint8_t len);
+/* Read len bytes from I2C slave. Returns true on success.
+ * len 支持 >255（7100 A7 大块读如 0x32=309B / 0x177=378B，须单事务一次读完）。 */
+bool i2c_7100_read(uint8_t addr, uint8_t *data, uint16_t len);
 
 /* Set I2C bus speed by bit-delay loop count.
  * Use I2C_7100_DELAY_FAST (25) when DSP is stopped.
